% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddexp.R
\name{ddexp}
\alias{ddexp}
\title{Função Densidade de Probabilidade da Exponencial Dupla (Laplace)}
\usage{
ddexp(x, mu, scale)
}
\arguments{
\item{x}{valor a ser avaliado.}

\item{mu}{parâmetro de locação.}

\item{scale}{parâmetro de escala.}
}
\description{
Função Densidade de Probabilidade da Exponencial Dupla (Laplace)
}
\details{
Seja \eqn{X} uma variável aleatória com distribuição de Laplace, então, sua
função densidade de probabilidade é dada por:
\deqn{f(x) = \dfrac{1}{2\beta} \exp\left\{ -\dfrac{|x-\mu|}{\beta} \right\}}
}
