% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desagrega.R
\name{desagrega}
\alias{desagrega}
\title{Desagregador Dados Funcionais}
\usage{
desagrega(
  data,
  y,
  policy = "sure",
  filter.number = 10,
  family = "DaubExPhase",
  a,
  s,
  tau
)
}
\arguments{
\item{data}{matriz contendo, em cada coluna, uma observações.}

\item{y}{matriz com os pesos conhecidos de cada funcional.}

\item{policy}{política para escolha de limiar. Os possíveis valores são
"sure", "universal", "cv", "fdr", "logistica", "epanechnikov". Para mais
detalhes, ver \code{\link[wavethresh]{wavethresh::threshold.wd}}.}

\item{filter.number}{controla a suavidade da ondaleta.}

\item{family}{Especifíca a família da ondaleta, e.g. "DaubExPhase",
"DaubLeAsymm".}

\item{a}{parâmetro \eqn{\alpha} da mistura.}

\item{s}{desvio padrão dos coeficientes de ondaleta, normalmente estimado
pelo MAD (median absolute deviation).}

\item{tau}{parâmetro \eqn{\tau} da logística.}
}
\value{
Retorna uma matriz com a função recuperar no domínio do tempo de
forma que cada coluna representa uma função recuperada.
}
\description{
Desagregador Dados Funcionais
}
\examples{
set.seed(282829)
x <- (1:1024)/1024
alpha_comp <- matrix(c(f_test()$bumps, f_test()$doppler), ncol=2)
sample <- sample_gen(alpha_comp, snr=5, I=10)
plot(sample)
fun_recup <- desagrega(sample$fun, sample$y)  # recuperando funções

# plot
par(mar=c(3, 4, 2, 2), mgp=c(2, 0.5, 0)); layout(matrix(c(1,1,2,3), 2, byrow=T))
plot(sample)
plot(x, fun_recup[,1], main='Bumps', ylab='y', col='blue', type='l')
lines(x, alpha_comp[,1])
legend('topright', lwd=2, bty='n', cex=0.85, col=c('black', 'blue'),
       legend=c('Função Verdadeira', 'Função Recuperada'))
plot(x, fun_recup[,2], main='Doppler', ylab='y', col='blue', type='l')
lines(x, alpha_comp[,2])
legend('bottomright', lwd=2, bty='n', cex=0.85, col=c('black', 'blue'),
       legend=c('Função Verdadeira', 'Função Recuperada'))
}
\references{
Sousa, A.R.S. (2024). A wavelet-based method in aggregated
functional data analysis. \emph{Monte Carlo Methods and Applications}, 30(1),
19-30. DOI: \href{https://doi.org/10.1515/mcma-2023-2016}{10.1515/mcma-2023-2016}.
}
