% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_test.R
\name{f_test}
\alias{f_test}
\title{Funções para Teste}
\usage{
f_test(n=1024, signal=7, snr=7, noise=FALSE)
}
\arguments{
\item{n}{números de pontos considerados da função.}

\item{signal}{desvio padrão da função verdadeira.}

\item{snr}{razão sinal-ruído.}

\item{noise}{indica se os pontos gerados devem apresentar um ruido gaussiano.}
}
\value{
Retorna uma lista com \code{n} pontos de cada função.
}
\description{
Funções para Teste
}
\details{
Todas as funções são normalizadas para apresentarem o mesmo desvio padrão
\code{signal}, cujo default é \eqn{7}.
}
\references{
Donoho, D.L. and Johnstone, I.M. (1994), Ideal spatial adaptation by wavelet
shrinkage. \emph{Biometrika}, 81, 425–455.
}
