% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_ponto.R
\name{gera_ponto}
\alias{gera_ponto}
\title{Gera um Ponto no Suprote da Posteriori com Erro Gamma}
\usage{
gera_ponto(n, y, lim_sup = 20, filter.number = 5, family = "DaubExPhase")
}
\arguments{
\item{n}{Quantidade de pontos por função.}

\item{y}{Dados.}

\item{lim_sup}{Valor máximo gerado para o ponto gerado a.}

\item{...}{configurações referênte à DWT, como \code{filter.number} e \code{family}.}
}
\description{
Gera um Ponto no Suprote da Posteriori com Erro Gamma
}
