% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logis_shrink_parallel.R
\name{logis_shrink_parallel}
\alias{logis_shrink_parallel}
\title{Regra de Encolhimento Bayesiana com Priori Logística}
\usage{
logis_shrink_parallel(d, a, s, t)
}
\arguments{
\item{d}{vetor de coeficientes de ondaleta empírico.}

\item{a}{parâmetro \eqn{\alpha} da mistura.}

\item{s}{desvio padrão dos coeficientes de ondaleta.}

\item{t}{parâmetro \eqn{\tau} da logística.}
}
\description{
Regra de Encolhimento Bayesiana com Priori Logística
}
