% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram}
\alias{ram}
\title{Gera Amostra da Posteriori com Erro Gamma}
\usage{
ram(
  theta_1 = NULL,
  S_1 = NULL,
  f,
  alpha = 0.8,
  tau = 2,
  beta,
  lambda,
  n_ite = 50000,
  gamma = 2/3,
  filter.number = 5,
  family = "DaubExPhase"
)
}
\description{
Gera Amostra da Posteriori com Erro Gamma
}
