% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu}
\alias{simu}
\title{Simulação para Política de Limiar}
\usage{
simu(fun_comp, rep, n=10, snr, policy='sure', filter.number=10)
}
\arguments{
\item{fun_comp}{objeto com as \eqn{L} funções componentes. Cada linha é uma
função.}

\item{rep}{quantidade de replicações da simulação.}

\item{n}{tamanho da amostra gerada.}

\item{snr}{razão sinal-ruído.}

\item{policy}{política para escolha de limiar. Os possíveis valores são "sure",
"universal", "cv", "fdr", "logistica", "epanechnikov". Para maide detalhes, ver
\code{\link[wavethresh]{wavethresh::threshold.wd}}.}

\item{filter.number}{controla a suavidade da ondaleta.}
}
\value{
Retorna um objeto da classe \code{data.frame} com o cálculo do MSE de cada
função e o AMSE.
}
\description{
Simulação para Política de Limiar
}
\details{
A função assume que os dados de entrada tem o desvio padrão 7 para efetuar o
cálculo da razão sinal-ruído. Além disso, os pesos para cada função são gerados
aleatóriamente seguindo uma \eqn{\mathcal{U}(0,1)}.
}
\examples{
bumps <- f_test()$bumps
doppler <- f_test()$doppler
par(mfrow=c(2,1))
plot(bumps, type='l', main='Bumps'); plot(doppler, type='l', main='Doppler')

# plan(multisession, workers=3)  # paralelizando execução
fun_comp <- matrix(c(bumps, doppler), nrow=2, byrow=T)
simu(fun_comp, rep=4, n=10, snr=5)
}
\references{
Sousa, A.R.S. (2024). A wavelet-based method in aggregated functional data
analysis. \emph{Monte Carlo Methods and Applications}, 30(1), 19-30.
\href{https://doi.org/10.1515/mcma-2023-2016}{doi:10.1515/mcma-2023-2016}.
}
