% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu2.R
\name{simu2}
\alias{simu2}
\title{DEPRECIADO: Simulação2 para Política de Limiar}
\usage{
simu2(fun_comp, rep, n = 10, snr, policy = "sure", filter.number = 10)
}
\description{
DEPRECIADO: Simulação2 para Política de Limiar
}
