% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_ram_gamma.R
\name{simu_ram_gamma}
\alias{simu_ram_gamma}
\title{Função de Simulação para Dados Funcionais com Erro Gamma}
\usage{
simu_ram_gamma(
  alpha_comp,
  I,
  n_ite,
  beta,
  lambda,
  alpha,
  tau,
  gamma = 2/3,
  filter.number = 5,
  family = "DaubExPhase",
  lim_sup = 15,
  bt = NULL
)
}
\arguments{
\item{alpha_comp}{Matriz com as funções componentes, cada coluna é uma
função.}

\item{I}{Tamanho da amostra gerada, isto é, quantidade de observações.}

\item{n_ite}{número de iterações (tamanho da cadeia gerada).}

\item{beta, lambda}{parâmetros do erro \eqn{Gamma(\beta, \lambda)}.}

\item{alpha}{Parâmetro da mistura da priori spike and slab utilizada.}

\item{tau}{Parâmetro da logística presente na priori.}

\item{gamma}{Parâmetro do algorítimo RAM. Por padrão, utiliza-se
\eqn{\frac{2}{3}}.}

\item{filter.number}{This selects the smoothness of wavelet that you want to use in the decomposition. By default this is 10, the Daubechies least-asymmetric orthonormal compactly supported wavelet with 10 vanishing moments. 

For the ``wavelets on the interval'' (\code{bc="interval"}) transform the filter number ranges from 1 to 8. See the table of filter coefficients indexed after the reference to Cohen, Daubechies and Vial, 1993.}

\item{family}{specifies the family of wavelets that you want to use. Two popular options are "DaubExPhase" and "DaubLeAsymm" but see the help for \code{\link[wavethresh]{filter.select}} for more possibilities.
 
This argument is ignored for the ``wavelets on the interval'' transform (\code{bc="interval"}).

Note that, as of version 4.6.1 you can use the Lina-Mayrand complex-valued
wavelets.
}

\item{bt}{Vetor contendo os índices de burn-in e thinning. Se não for
especificado será utilizado uma queima de 5\% das observações e um thinning
de 3 observações.}
}
\description{
Função de Simulação para Dados Funcionais com Erro Gamma
}
