% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_test.R
\name{f_test}
\alias{f_test}
\title{Funções para Teste}
\usage{
f_test(n = 1024, signal = 7, snr = 7, noise = FALSE)
}
\arguments{
\item{n}{números de pontos considerados da função.}

\item{signal}{desvio padrão da função verdadeira.}

\item{snr}{razão sinal-ruído.}

\item{noise}{indica se os pontos gerados devem apresentar um ruido gaussiano.}
}
\description{
Funções para Teste
}
\references{
Donoho, D.L. and Johnstone, I.M. (1994), Ideal spatial adaptation by wavelet
shrinkage. \enph{Biometrika}, 81, 425–455.

Sousa, A.R.S. (2020). Bayesian wavelet shrinkage with logistic prior.
\emph{Communications in Statistics - Simulation and Computation}.
}
