% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_ponto.R
\name{gera_ponto}
\alias{gera_ponto}
\title{Gera um Ponto no Suprote da Posteriori com Erro Gamma}
\usage{
gera_ponto(y, lim_sup = 20, filter.number = 5, family = "DaubExPhase")
}
\arguments{
\item{y}{Valores observados.}

\item{lim_sup}{Valor máximo para uma coordenada gerado para o ponto gerado a.}

\item{filter.number}{This selects the smoothness of wavelet that you want to use in the decomposition. By default this is 10, the Daubechies least-asymmetric orthonormal compactly supported wavelet with 10 vanishing moments. 

For the ``wavelets on the interval'' (\code{bc="interval"}) transform the filter number ranges from 1 to 8. See the table of filter coefficients indexed after the reference to Cohen, Daubechies and Vial, 1993.}

\item{family}{specifies the family of wavelets that you want to use. Two popular options are "DaubExPhase" and "DaubLeAsymm" but see the help for \code{\link[wavethresh]{filter.select}} for more possibilities.
 
This argument is ignored for the ``wavelets on the interval'' transform (\code{bc="interval"}).

Note that, as of version 4.6.1 you can use the Lina-Mayrand complex-valued
wavelets.
}
}
\description{
Gera um ponto no domínio do tempo da função original garantindo que esta
dentro do suporte da posteriori com erro Gamma \code{\link[=post_gamma]{post_gamma()}}.
}
\examples{
y <- c(1,2,7,2,5,6,0,3)  # valores observados
d <- wd(y, filter.number = 5, family = 'DaubExPhase')
(theta <- gera_ponto(y))
post_gamma(theta, d, beta = 5, lambda = 1, tau = 4, alpha = 0.7)  # posteriori
}
