% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_gamma.R
\name{post_gamma}
\alias{post_gamma}
\title{Posteriori dos Coeficientes de Ondaletas Considerando Erro Gamma}
\usage{
post_gamma(
  theta,
  d,
  beta,
  lambda,
  tau,
  alpha = 0.8,
  filter.number = 5,
  family = "DaubExPhase",
  suppressMessage = T
)
}
\arguments{
\item{theta}{vetor}

\item{d}{coeficientes empíricos.}

\item{beta, lambda}{parâmetro da \eqn{Gamma(\beta, \lambda)}.}

\item{tau}{parâmetro da logística.}

\item{alpha}{coeficiente da spike and slab (0.8 é o valor padrão).}

\item{filter.number}{This selects the smoothness of wavelet that you want to use in the decomposition. By default this is 10, the Daubechies least-asymmetric orthonormal compactly supported wavelet with 10 vanishing moments. 

For the ``wavelets on the interval'' (\code{bc="interval"}) transform the filter number ranges from 1 to 8. See the table of filter coefficients indexed after the reference to Cohen, Daubechies and Vial, 1993.}

\item{family}{specifies the family of wavelets that you want to use. Two popular options are "DaubExPhase" and "DaubLeAsymm" but see the help for \code{\link[wavethresh]{filter.select}} for more possibilities.
 
This argument is ignored for the ``wavelets on the interval'' transform (\code{bc="interval"}).

Note that, as of version 4.6.1 you can use the Lina-Mayrand complex-valued
wavelets.
}

\item{suppressMessage}{Se \code{TRUE} (default) suprime a mensagem "Fora do
suporte." caso o ponto \code{theta} esteja fora do domínio,}
}
\description{
Posteriori dos Coeficientes de Ondaletas Considerando Erro Gamma
}
\details{
Partindo do modelo vetorial no dominio do tempo
\deqn{
\boldsymbol y = \boldsymbol f + \boldsymbol e
}
com \eqn{e_i \sim Gamma(\beta, \lambda)}. Então, no domínio das ondaletas,
temos que o modelo é dado por
\deqn{
\boldsymbol d = \boldsymbol\theta + \boldsymbol\varepsilon}
onde \eqn{\boldsymbol d = W \boldsymbol y} são os coeficiente empíricos,
\eqn{\boldsymbol\theta = W \boldsymbol f} são os coeficientes de ondaleta
verdadeiros e \eqn{\boldsymbol \varepsilon = W \boldsymbol e} é o erro no
domínio das ondaletas Além disso, atribuindo a priori dos coeficientes de
ondaleta \eqn{\boldsymbol\theta} consiederado que seus elementos são
independentes, tem-se:
\deqn{
\displaystyle\pi(\boldsymbol\theta; \alpha, \tau) = \prod_{i=1}^n[\alpha
\delta_0(\theta_i) + (1 - \alpha) g(\theta_i; \tau)]
}
onde \eqn{\alpha \in (0,1)}, \eqn{\delta_0} é o delta de Dirac com massa em
0 e \eqn{g(\theta; \tau)} é a função densidade de probabilidade logística,
dada por
\deqn{
g(\theta_i; \tau) = \dfrac{\exp\left\{-\frac{\theta}{\tau}\right\}}{\tau
\left( 1 + \exp\left\{-\frac{\theta}{\tau}\right\} \right)^2} \;
\mathcal{I}_\mathbb{R}(\theta) \;, \hspace{0.45cm} \tau > 0
}
Portanto, com isso é possível obter a verossimilhança e computar a
posteriori, dada por
\deqn{
\displaystyle \pi(\boldsymbol\theta \mid \boldsymbol d) \propto \prod_{i=1}^n
\left[ \alpha \delta_0(\theta_i) + (1 - \alpha)
\frac{\exp\left\{-\frac{\theta_i}{\tau}\right\}}{\tau \left( 1 +
\exp\left\{-\frac{\theta_i}{\tau}\right\} \right)^2} \right] \\ \times
\exp\left\{-\lambda \sum_{i=1}^n\sum_{k=1}^n w_{ki} (d_k -\theta_k)\right\}
\left( \prod_{i=1}^n\sum_{k=1}^n w_{ki} (d_k -\theta_k) \right)^{\beta-1} \\
\times \prod_{i=1}^n \mathcal{I}_{(0,\infty)}\left(\sum_{k=1}^n w_{ki} (d_k
-\theta_k)\right)
}
}
\references{
Sousa, A.R.S., Garcia, N.L (2023): Wavelet shrinkage in
nonparametric regression models with positive noise. \emph{Journal of
Statistical Computation and Simulation}, DOI:
\href{https://doi.org/10.1080/00949655.2023.2215372}{10.1080/00949655.2023.2215372}.
}
