% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_gamma.R
\name{ram_gamma}
\alias{ram_gamma}
\title{Gera Amostra da Posteriori com Erro Gamma por Meio do Algorítimo RAM}
\usage{
ram_gamma(
  theta_1 = NULL,
  S_1 = NULL,
  f,
  alpha = 0.8,
  tau = 2,
  beta,
  lambda,
  n_ite = 50000,
  gamma = 2/3,
  filter.number = 5,
  family = "DaubExPhase"
)
}
\arguments{
\item{theta_1}{Chute inicial para \eqn{\theta}. Se não for definido, será
gerado um ponto aleatório baseado na função \code{gera_ponto}.}

\item{S_1}{Chute inicial para a matriz S. Se não for definida será utilziado
a identidaide.}

\item{f}{Valores verdadeiros.}

\item{alpha}{Parâmetro da mistura da priori spike and slab utilizada.}

\item{tau}{Parâmetro da logística presente na priori.}

\item{beta, lambda}{parâmetros do erro \eqn{Gamma(\beta, \lambda)}.}

\item{n_ite}{número de iterações.}

\item{gamma}{Parâmetro do algorítimo RAM.}

\item{filter.number}{This selects the smoothness of wavelet that you want to use in the decomposition. By default this is 10, the Daubechies least-asymmetric orthonormal compactly supported wavelet with 10 vanishing moments. 

For the ``wavelets on the interval'' (\code{bc="interval"}) transform the filter number ranges from 1 to 8. See the table of filter coefficients indexed after the reference to Cohen, Daubechies and Vial, 1993.}

\item{family}{specifies the family of wavelets that you want to use. Two popular options are "DaubExPhase" and "DaubLeAsymm" but see the help for \code{\link[wavethresh]{filter.select}} for more possibilities.
 
This argument is ignored for the ``wavelets on the interval'' transform (\code{bc="interval"}).

Note that, as of version 4.6.1 you can use the Lina-Mayrand complex-valued
wavelets.
}
}
\description{
Gera Amostra da Posteriori com Erro Gamma por Meio do Algorítimo RAM
}
\examples{
sample <- ram_gamma(f = f_test(n = 8)$bumps, beta = 10, lambda = 30, n_ite = 3000)
colMeans(sample$theta)
}
\references{
Vihola, M. (2012). Robust adaptive Metropolis algorithm with coerced
acceptance rate. Stat Comput 22, 997–1008. DOI:
\href{https://doi.org/10.1007/s11222-011-9269-5}{10.1007/s11222-011-9269-5}.

Sousa, A.R.S., Garcia, N.L. (2023). Wavelet shrinkage in
nonparametric regression models with positive noise. \emph{Journal of
Statistical Computation and Simulation}. DOI:
\href{https://doi.org/10.1080/00949655.2023.2215372}{10.1080/00949655.2023.2215372}.
}
