% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gen.R
\name{sample_gen}
\alias{sample_gen}
\title{Gera uma Amostra de Dados Funcionais}
\usage{
sample_gen(
  fun_comp,
  snr,
  n = 10,
  signal = 7,
  beta,
  lambda,
  erro = "normal",
  stand = FALSE
)
}
\arguments{
\item{fun_comp}{objeto com as \eqn{L} funções componentes, onde cada linha é uma
função.}

\item{snr}{razão sinal-ruído.}

\item{n}{tamanho da amostra gerada.}

\item{signal}{desvio padrão dos dados após padronização.}

\item{beta, lambda}{Parâmetros para erro \eqn{Gamma(\beta, \lambda)}.}

\item{erro}{tipo de erro, podendo ser "normal" ou "gamma".}

\item{stand}{Se FALSE (default) as funções componentes não são padronizada para
terem razão sinal-ruído igual a \code{signal}.}
}
\value{
Retorna uma \code{lista} contendo dois elementos: \code{fun}, onde cada linha
é um elemento da amostra gerada, e \code{y}, os quais são os pesos
utilizados.
}
\description{
Esta função gera uma amostra de dados funcionais agregados baseado na entrada
das funções componentes, razão sinal-ruído e tamanho da amostra.
}
\examples{
bumps <- f_test()$bumps; doppler <- f_test()$doppler
par(mfrow=c(2,1))
plot(bumps, type='l', main='Bumps')
plot(doppler, type='l', main='Doppler')
par(mfrow=c(1,1))

fun_comp <- matrix(c(bumps, doppler), nrow=2, byrow=T)
sample <- sample_gen(fun_comp, snr=5, n=7)
plot(1:1024, main='Amostra Gerada', ylab='', type='n',
     ylim=c(min(sample$fun), max(sample$fun)))
lapply(1:nrow(sample$fun), \(i) lines(sample$fun[i,], col=i)) |> invisible()
}
