% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu}
\alias{simu}
\title{Simulação para Política de Limiar}
\usage{
simu(fun_comp, snr, rep, n=10, policy='sure', filter.number=10,
     stand=T, signal=7)
}
\arguments{
\item{fun_comp}{objeto com as \eqn{L} funções componentes, onde cada linha é uma
função.}

\item{snr}{razão sinal-ruído.}

\item{rep}{quantidade de replicações da simulação.}

\item{n}{tamanho da amostra gerada.}

\item{policy}{política para escolha de limiar. Os possíveis valores são "sure",
"universal", "cv", "fdr", "logistica", "epanechnikov". Para maide detalhes, ver
\code{\link[wavethresh]{wavethresh::threshold.wd}}.}

\item{filter.number}{controla a suavidade da ondaleta.}

\item{stand}{Se FALSE (default) as funções componentes não são padronizada para
terem razão sinal-ruído igual a \code{signal}.}

\item{signal}{desvio padrão dos dados após padronização.}
}
\value{
Retorna um objeto da classe \code{data.frame} com o cálculo do MSE de cada função
e o AMSE.
}
\description{
Simulação para Política de Limiar
}
\details{
A função assume que, para o cálculo da razão sinal-ruído, que \eqn{sd(sinal)=7}.
Os pesos para cada função são gerados aleatóriamente seguindo uma
\eqn{\mathcal{U}(0,1)} e são modificados de forma a garantir que sua soma é
\eqn{1}.
}
\examples{
bumps <- f_test()$bumps; doppler <- f_test()$doppler
par(mfrow=c(2,1))
plot(bumps, type='l', main='Bumps'); plot(doppler, type='l', main='Doppler')

# plan(multisession, workers=2)  # paralelizando execução
fun_comp <- matrix(c(bumps, doppler), nrow=2, byrow=T)
simu(fun_comp, rep=4, n=10, snr=5)
}
\references{
Sousa, A.R.S. (2024). A wavelet-based method in aggregated functional data
analysis. \emph{Monte Carlo Methods and Applications}, 30(1), 19-30.
\href{https://doi.org/10.1515/mcma-2023-2016}{doi:10.1515/mcma-2023-2016}.
}
